requirements.txt
pytorch-lightning==1.2.3
pathlib2==2.3.5
glob2==0.7
h5py==2.10.0
jupyter==1.0.0
lifelines==0.26.0
opencv-python==4.2.0.34
openslide-python==1.1.2
#scikit-survival==0.14.0
scipy==1.5.4
tensorboard==2.4.0
tqdm==4.46.1
tifffile==2021.1.11
gpustat
torchsummary==1.5.1
timm==0.3.2
pytorch-toolbelt==0.4.0
addict==2.2.1
pycox==0.2.2
loguru==0.5.3
einops==0.3.0
nystrom-attention==0.0.9
torchmetrics==0.4.1
torch-optimizer==0.0.1a13
pandas==1.2.3
nmslib==2.0.6
tensorboardX==2.0
